Moved factory.py from ctapipe/calib/camera to ctapipe/core
