+++ ctapipe/calib/camera/charge_extractors.py
@@ -1,9 +1,10 @@
 from abc import abstractmethod
+
 import numpy as np
 from traitlets import Int, CaselessStrEnum
 
+from core.factory import Factory
 from ctapipe.core import Component
-from ctapipe.calib.camera.factory import Factory
 
 
 class ChargeExtractor(Component):
