+++ ctapipe/core/factory.py
